fix: rename package main to internal
entropy.go and conversions.go had package main attached
