+++ internal/conversions.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 /*
 * StringArrayToOrdinal - converts an array of strings to an array of integers
